LayoutTests/platform/qt/fast/events/continuous-platform-wheelevent-in-scrolling-div-expected.txt

FAIL event.wheelDeltaY should be -600. Was 0.
FAIL event.wheelDeltaX should be -300. Was -12000.
FAIL event.wheelDelta should be -300. Was -12000.
FAIL event.wheelDeltaY should be -600. Was -24000.
FAIL event.wheelDeltaX should be -300. Was 0.
FAIL event.wheelDelta should be -600. Was -24000.
PASS div.scrollTop is window.expectedScrollTop
PASS div.scrollLeft is window.expectedScrollLeft
